refactor: conditional escrow modular structure
